backend/requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
fastapi
uvicorn
pydantic
python-dotenv
requests